fixes refactoring; enhances description
