+++ mescal/energy_data_handling/area_accounting/area_variable_sum_calculator.py
@@ -1,6 +1,6 @@
 import pandas as pd
 
-from mescal.energy_data_handling.area_border.area_variable_base import AreaVariableCalculatorBase
+from mescal.energy_data_handling.area_accounting.area_variable_base import AreaVariableCalculatorBase
 
 
 class AreaSumCalculator(AreaVariableCalculatorBase):
