+++ mescal/energy_data_handling/area_accounting/border_variable_capacity.py
@@ -5,7 +5,7 @@ import numpy as np
 import pandas as pd
 
 from mescal.energy_data_handling.network_lines_data import NetworkLineCapacitiesData
-from mescal.energy_data_handling.area_border.border_variable_base import AreaBorderVariableCalculatorBase
+from mescal.energy_data_handling.area_accounting.border_variable_base import AreaBorderVariableCalculatorBase
 
 
 class BorderCapacityCalculator(AreaBorderVariableCalculatorBase):
